requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pymongo==4.7.2
redis==5.0.4
kafka-python==2.0.2
python-multipart==0.0.9
pandas==2.2.1
streamlit==1.32.2
plotly==5.20.0
